Only try to mount devices with "usb" in their paths
We're only interested in USB keys, so filtering for USB devices will
reduce the work we have to do.  Since we don't get any information about
the device's bus through netlink, a hacky but effective method of doing
this is to check that "usb" features in the device's sysfs path.

Change-Id: I508ebaef61f3b44b9d9895874d0c788de58f5e19